+++ sdk/storage/azure_storage_blob/src/lib.rs
@@ -7,6 +7,8 @@
 #![allow(dead_code)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
+#[cfg(feature = "arrow")]
+mod arrow;
 pub(crate) mod buffers;
 pub mod clients;
 #[allow(unused_imports)]